refactor(direct-control): extract component id primitive
Move the Civ7 ComponentID schema, type guard, and assertion helper into packages/civ7-direct-control/src/civ7-component-id.ts, and move Civ7DirectControlError ownership into packages/civ7-direct-control/src/direct-control-error.ts so the primitive can throw the package error without a facade cycle. Re-export both owners through src/index.ts to preserve the public package API.

Update the direct-control build script so tsup emits JS and tsc emits declarations for the new multi-file source layout. Record the OpenSpec source-owner progress and preserve the protected direct-control modularization TODO stash as compatible evidence: tests came first, then a named primitive/error owner move.

Validation: bun run --cwd packages/civ7-direct-control test -- public-api.test.ts session.test.ts restart-lifecycle.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
