+++ packages/civ7-direct-control/src/direct-control-error.ts
@@ -1,0 +1,38 @@
+export type Civ7DirectControlErrorCode =
+  | "invalid-port"
+  | "connection-timeout"
+  | "connection-failed"
+  | "response-timeout"
+  | "socket-closed"
+  | "state-not-found"
+  | "no-hosts"
+  | "all-hosts-unavailable"
+  | "command-failed"
+  | "log-timeout"
+  | "setup-api-unavailable"
+  | "setup-phase-invalid"
+  | "setup-map-row-missing"
+  | "setup-parameter-invalid"
+  | "setup-apply-timeout"
+  | "setup-readback-mismatch"
+  | "setup-start-timeout"
+  | "setup-seed-mismatch"
+  | "setup-map-size-mismatch"
+  | "setup-config-load-failed"
+  | "setup-config-proof-missing";
+
+export class Civ7DirectControlError extends Error {
+  readonly code: Civ7DirectControlErrorCode;
+  readonly details?: unknown;
+
+  constructor(
+    code: Civ7DirectControlErrorCode,
+    message: string,
+    options?: { cause?: unknown; details?: unknown },
+  ) {
+    super(message, options?.cause !== undefined ? { cause: options.cause } : undefined);
+    this.name = "Civ7DirectControlError";
+    this.code = code;
+    this.details = options?.details;
+  }
+}
